requirements.txt
cython
scipy
shapely
timm
h5py
submitit
scikit-image
wandb
setuptools
numpy==1.22.4
Pillow==9.0.1
pycocotools~=2.0.4
fvcore
tabulate
tqdm
ftfy
regex
opencv-python
open_clip_torch==2.16.0